splitting analytic events file
