+++ app/client/src/ee/utils/analyticsUtilTypes.ts
@@ -1,0 +1,1 @@
+export * from "ce/utils/analyticsUtilTypes";
